refactor: 로그 관련 기능을 Support 하위로 이동
