+++ backend/src/main/java/com/woowacourse/momo/global/exception/ControllerAdvice.java
@@ -10,7 +10,7 @@ import com.woowacourse.momo.global.exception.dto.response.ExceptionResponse;
 import com.woowacourse.momo.global.exception.exception.ErrorCode;
 import com.woowacourse.momo.global.exception.exception.GlobalErrorCode;
 import com.woowacourse.momo.global.exception.exception.MomoException;
-import com.woowacourse.momo.global.logging.UnhandledErrorLogging;
+import com.woowacourse.momo.support.logging.UnhandledErrorLogging;
 
 @RestControllerAdvice
 public class ControllerAdvice {
