+++ backend/src/main/java/com/woowacourse/momo/support/logging/ConsolePrettier.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.global.logging;
+package com.woowacourse.momo.support.logging;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
