+++ backend/src/main/java/com/woowacourse/momo/support/logging/TraceExtractor.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.global.logging;
+package com.woowacourse.momo.support.logging;
 
 import java.io.IOException;
 import java.io.PrintWriter;
