+++ backend/src/main/java/com/woowacourse/momo/support/logging/exception/LogException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.global.logging.exception;
+package com.woowacourse.momo.support.logging.exception;
 
 public class LogException extends RuntimeException {
